requirements.txt
# End-to-End ML Workflow with Observability - Requirements
# Customer Churn Detection Pipeline Dependencies

# Core ML and Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0

# Snowflake Libraries
snowflake-snowpark-python>=1.11.0
snowflake-ml-python>=1.2.0

# Optional: Streamlit for UI (if needed)
streamlit>=1.28.0

# Logging and Configuration
python-dotenv>=1.0.0

# Development and Testing Dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Jupyter for notebook development (optional)
jupyter>=1.0.0
ipykernel>=6.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.0.0
